feat: expose testcase API in library
