+++ Cargo.toml
@@ -1,89 +1,89 @@
 [package]
-name = "varlociraptor"
-version = "5.3.3"
 authors = ["Johannes Köster <johannes.koester@tu-dortmund.de>", "David Lähnemann <david.laehnemann@uni-duesseldorf.de>"]
 description = "A library for calling of genomic variants using a latent variable model."
-repository = "https://github.com/varlociraptor/varlociraptor"
-readme = "README.md"
-license = "GPL-3.0"
-include = ["src/**/*", "Cargo.toml", "CHANGELOG.md", "README.md", "LICENSE", "templates/*"]
 edition = "2018"
+include = ["src/**/*", "Cargo.toml", "CHANGELOG.md", "README.md", "LICENSE", "templates/*"]
+license = "GPL-3.0"
+name = "varlociraptor"
+readme = "README.md"
+repository = "https://github.com/varlociraptor/varlociraptor"
+version = "5.3.3"
 
 [dependencies]
-bio = "0.40.0"
-bio-types = { version = ">=0.12", features = ["serde"] }
-rust-htslib = "0.40.2"
-log = "0.4"
 GSL = "6.0"
+anyhow = "1.0"
+approx = "0.5.0"
+askama = "0.10.5"
+auto_ops = "0.3.0"
+bincode = "1.3"
+bio = "0.40.0"
+bio-types = {version = ">=0.12", features = ["serde"]}
+boolean_expression = "0.4"
+bv = "0.11"
+byteorder = "1.3"
+cached = "0.26"
+clap = {version = "2", features = ["yaml", "color", "suggestions"]}
+counter = "0.5"
+csv = "1"
+data-encoding = "2.3"
+derefable = "0.1"
+derive-new = "0.5"
+derive_builder = "0.11.2"
+eval = "0.4"
+fern = "0.6"
+getset = "0.1.0"
+half = {version = "1.4", features = ["serde"]}
 itertools = "0.10"
 itertools-num = "0.1"
-approx = "0.5.0"
-rusty-machine="0.5"
-ordered-float = { version = "1", features = ["serde"] }
+jemallocator = "0.3.2"
+lazy_static = "1"
+log = "0.4"
+lru = "0.7"
+lru_time_cache = "0.11"
 ndarray = "0.15"
-vec_map = "0.8"
+num-traits = "0.2"
+ordered-float = {version = "1", features = ["serde"]}
+paste = "1"
+pest = "2"
+pest_derive = "2"
+petgraph = "0.6"
+progress-logger = "0.3"
+rand = "0.8"
+rayon = "1"
 regex = "1"
+ring = "0.16"
+rust-htslib = "0.40.2"
+rusty-machine = "0.5"
 serde = "1"
 serde_derive = "1"
 serde_json = "1"
 serde_yaml = ">=0.8.4,<0.9"
-csv = "1"
-lazy_static = "1"
-cached = "0.26"
+shrinkwraprs = "0.3.0"
 statrs = "0.15"
-clap = {version = "2", features = ["yaml", "color", "suggestions"]}
 structopt = "0.3"
 strum = ">= 0.16, < 0.24"
 strum_macros = ">= 0.16, < 0.24"
-rand = "0.8"
-derive_builder = "0.11.2"
-fern = "0.6"
-time = "0.3"
-counter = "0.5"
-askama = "0.10.5"
-yaml-rust = ">=0.4.1,<0.5"
 tempfile = "3"
-eval = "0.4"
-rayon = "1"
-petgraph = "0.6"
-derive-new = "0.5"
-pest = "2"
-pest_derive = "2"
-getset = "0.1.0"
-half = { version = "1.4", features = ["serde"] }
-bincode = "1.3"
-bv = "0.11"
-shrinkwraprs = "0.3.0"
-byteorder = "1.3"
 thiserror = "1.0"
-anyhow = "1.0"
-derefable = "0.1"
-lru_time_cache = "0.11"
-lru = "0.7"
+time = "0.3"
 typed-builder = "0.9"
-jemallocator = "0.3.2"
-ring = "0.16"
-data-encoding = "2.3"
-boolean_expression = "0.4"
-uuid = { version = "0.8", features = ["v4"] }
-auto_ops = "0.3.0"
-num-traits = "0.2"
-progress-logger = "0.3"
+uuid = {version = "0.8", features = ["v4"]}
+vec_map = "0.8"
+yaml-rust = ">=0.4.1,<0.5"
 
 [dev-dependencies]
 env_logger = "0.9.0"
 serde_json = "1"
-paste = "1"
 
 [features]
 default = []
 use-intrinsics = ["half/use-intrinsics"]
 
 [profile.release]
-lto = "fat"
 codegen-units = 1
+lto = "fat"
 
 [profile.develop]
-inherits="release"
-lto = false
 codegen-units = 128
+inherits = "release"
+lto = false
